fix: won't try to delete an avatar if a user doesn't already have an avatar

+++ frontend/src/components/profile/UserDetails.vue
@@ -246,9 +246,11 @@ async function avatarChanged (event) {
 
     await onSave("avatar");
 
-    await directus(`/files/${oldId}`, {
-      method: "DELETE"
-    });
+    if (oldId) {
+      await directus(`/files/${oldId}`, {
+        method: "DELETE"
+      });
+    }
   } finally {
     uploadingAvatar.value = false;
   }
